# src/SProfTIAPI/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "SProfTI API",
    "description" : "Seletor de Profissinais de TI API",
    "contact" : {
      "email" : "rubensagnelo@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/rubensagnelo/sprotiapi/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
...
  "paths" : {
...
    "/usuarioplataforma/{idplataforma}/{idUsuarioplataforma}" : {
      "get" : {
        "tags" : [ "usuarioplataforma" ],
        "summary" : "Busca um usuário de uma plataforma",
        "description" : "Busca um usuário de uma plataforma (Github, SteckOverflow)",
        "operationId" : "Buscarusuariopnaplataforma",
        "parameters" : [ {
          "name" : "idplataforma",
          "in" : "path",
          "description" : "id do usuário na plataforma (user_id no StackOverflow e id no GitHub)",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "idUsuarioplataforma",
          "in" : "path",
          "description" : "quantidades de registros",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "resultados da consulta",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/usuarioplataformaitem"
                }
              }
            }
          },
          "400" : {
            "description" : "valor do parametro inválido"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "usuarioplataformaitem" : {
        "required" : [ "idplataforma", "idusuarioplatafaorma", "nome" ],
        "type" : "object",
        "properties" : {
          "idplataforma" : {
            "type" : "integer",
            "description" : "Identificado do usuário na plataforma retornado pela API da plataforma",
            "example" : 123
          },
          "idusuarioplatafaorma" : {
            "type" : "string",
            "description" : "Identificador do usuário na plataforma (user_id no StackOverflow e id no GitHub)",
            "example" : "12345"
          },
          "nome" : {
            "type" : "string",
            "description" : "nome do usuáriodescrito na plataforma",
            "example" : "João da Silva"
          },
          "avatar" : {
            "type" : "string",
            "description" : "url da imagem do avatar  (profile_image no StackOverflow e avatar_url no GitHub",
            "example" : "https://github.com/images/error/octocat_happy.gif"
          },
          "email" : {
            "type" : "string",
            "description" : "e-mail do usuário da",
            "example" : "octocat@github.com"
          },
          "reputacao" : {
            "type" : "integer",
            "description" : "valor resultante de formula que cruza alguns dados do usuários nas respectivas plataformas",
            "example" : 2041
          }
        },
        "example" : {
          "idusuarioplatafaorma" : "12345",
          "idplataforma" : 123,
          "nome" : "João da Silva",
          "avatar" : "https://github.com/images/error/octocat_happy.gif",
          "reputacao" : 2041,
          "email" : "octocat@github.com"
        }
      },
...
    }
  }
}